refactor(datasets): Un-nest datasets
Make datasets directly importable from the directory.
